+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 - #1148 Add e2e test for DAYOFWEEK
 - #1130 Infer hive folder partition
 - #1139 Adding centralized task executor for kernels
+- #1188 Implement upper/lower operators
+
 
 ## Improvements
 - #878 Adding calcite rule for window functions. (Window functions not supported yet)